+++ Overlapp.Shared/Model/Domain/CreditAggregate.cs
@@ -6,5 +6,5 @@ using System.Threading.Tasks;
 
 namespace Overlapp.Shared.Model
 {
-	public record CreditAggregate(string Name, string CharacterOrJob, string Department, long Id, string CreditId, int InstanceCount, float Popularity, string? Image, IMediaRecord? Item);
+	public record CreditAggregate(string Name, string CharacterOrJob, string Department, int Id, string CreditId, int InstanceCount, float Popularity, string? Image, IMediaRecord? Item);
 }
